Erstellt FeatureVector nach angepasstem Schema, für einen Ordner als .dat-Datei

+++ src/de/hszg/learner/FeatureVector.java
@@ -19,6 +19,6 @@ public interface FeatureVector extends Serializable{
 	 * @param i the index of the feature in the vector
 	 * @return the value of the feater with index i
 	 */
-	int getFeatureValue(int i);
+	double getFeatureValue(int i);
 
 }
